✨ Support multiple rows and full auto layouting
The PDF generation will now - unless configured
otherwise - try to make the best of the available
space and generate multiple rows of cards. It
will also detect whether a horizontal or vertical
fold line is best to get as many as possible cards
on one page.

All of this can of course be turned off by deselecting
"Auto" for foldline and unchecking "allow multiple rows".